backend/requirements.txt
blinker==1.9.0
click==8.3.2
Flask==3.1.3
flask-cors==6.0.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pillow==12.2.0
pypdf==6.10.1
Werkzeug==3.1.8
